# openapi.yaml
openapi: 3.1.0
info:
  title: Sportsdeck API
  version: 1.0.0
  description: API documentation for Sportsdeck, including endpoints for authentication, users, matches, forums, polls, and moderation.
servers:
  - url: /api
    description: Base API endpoint
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /users/{userId}/posts:
    get:
      summary: Get user threads and posts
      description: Recent threads & posts by user.
      security: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          example: "clq9s8r7k0000abc1def2ghi3"
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Threads and posts data
          content:
            application/json:
              example:
                threads:
                  data: []
                  pagination:
                    page: 1
                    limit: 20
                    totalItems: 0
                    totalPages: 1
                posts:
                  data: []
                  pagination:
                    page: 1
                    limit: 20
                    totalItems: 0
                    totalPages: 1
        '404':
          description: User not found
        '500':
          description: Fetch failure
...
  /users/me:
    get:
      summary: Get own profile
      description: Fetch own profile fields for the authenticated user.
      responses:
        '200':
          description: Profile object
          content:
            application/json:
              example:
                id: "clq9s8r7k0000abc1def2ghi3"
                username: "johndoe"
                email: "user@example.com"
        '404':
          description: User not found
        '500':
          description: Load failure

    patch:
      summary: Update own profile
      description: Update username or favoriteTeamId.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                favoriteTeamId:
                  type: string
            example:
              username: "newjohndoe"
              favoriteTeamId: "team123"
      responses:
        '200':
          description: Updated user profile
        '500':
          description: Update failed
...
  /users/me/followers/{followerId}:
    delete:
      summary: Remove a follower
      description: Remove specific follower from your list.
      parameters:
        - in: path
          name: followerId
          required: true
          schema:
            type: string
          example: "follower123"
      responses:
        '200':
          description: Success
        '400':
          description: Invalid or not a follower
        '404':
          description: Follower not found
        '500':
          description: Removal failed
...